# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Manage PDF documents",
        "description": "Simple API for manage PDF files",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/documents/": {
            "get": {
                "summary": "List documents",
                "description": "Returns a list of all documents registered",
                "operationId": "list_documents",
                "responses": {
                    "200": {
                        "description": "Showing all Documents"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/documents": {
            "post": {
                "summary": "create a document",
                "description": "create a document",
                "operationId": "new_document",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DocumentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "document created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocumentSchema"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "server error, try again"
                    }
                }
            }
        },
        "/api/documents/{id}": {
            "get": {
                "summary": "Get document",
                "description": "Returns information about document by id",
                "operationId": "get_document_by_id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "document retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocumentSchema"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "server error, try again"
                    }
                }
            },
            "put": {
                "summary": "update a document",
                "description": "update a document",
                "operationId": "update_document",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DocumentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "document updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocumentSchema"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "server error, try again"
                    }
                }
            },
            "delete": {
                "summary": "delete a document",
                "description": "delete a document",
                "operationId": "delete_document",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DocumentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "document deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocumentSchema"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "server error, try again"
                    }
                }
            }
        },
        "/api/documents/{id}/download": {
            "get": {
                "summary": "Download document",
                "description": "Returns pdf file for a document",
                "operationId": "download_document_by_id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "download successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocumentSchema"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "server error, try again"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DocumentRequest": {
                "properties": {
                    "id": {
                        "description": "id",
                        "type": "integer",
                        "example": "1",
                        "nullable": "false"
                    },
                    "name": {
                        "description": "name",
                        "type": "string",
                        "example": "string",
                        "nullable": "false"
                    },
                    "type_id": {
                        "description": "Type id",
                        "type": "integer",
                        "example": "1",
                        "nullable": "false"
                    }
                },
                "type": "object"
            },
...
            "DocumentSchema": {
                "properties": {
                    "id": {
                        "description": "id",
                        "type": "integer",
                        "example": "1",
                        "nullable": "false"
                    },
                    "name": {
                        "description": "name",
                        "type": "string",
                        "example": "string",
                        "nullable": "false"
                    },
                    "type_id": {
                        "description": "Type id",
                        "type": "integer",
                        "example": "1",
                        "nullable": "false"
                    }
                },
                "type": "object"
            },
...
        }
    }
}